nightguide-backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional ML libraries (uncomment when CNN model is ready)
# tensorflow>=2.13.0
# torch>=2.0.1
# torchvision>=0.15.2 